% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedata.R
\name{makedata}
\alias{makedata}
\title{Helper function to convert polygon to data frame}
\usage{
makedata(data, thin = T, unlist = T, tolerance = 0.1)
}
\description{
Helper function to convert polygon to data frame
}
