% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ozbig}
\alias{ozbig}
\title{ozbig}
\format{A MULTIPOLYGON with 11 variables
\describe{
  ...
}}
\usage{
ozbig
}
\description{
A dataset containing geometry for Australia and territories
}
\keyword{datasets}
