% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team1.R
\name{team1}
\alias{team1}
\title{lab3 work with team 1 function}
\usage{
team1(file, tolerance)
}
\arguments{
\item{tolerance}{}
}
\value{
plot
}
\description{
lab3 work with team 1 function
}
\examples{
team1(file="./data/gadm36_AUS_shp/gadm36_AUS_1.shp", tolerance=0.1)
}
