% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_12.R
\name{team_12}
\alias{team_12}
\title{Get polygon and geographic information from a shapefile that is thinned to a more feasible dimension. This is team 12's results}
\usage{
team_12(file, tolerance = 0.1, fileread = TRUE)
}
\arguments{
\item{file}{path to a shapefile (.shp) with spatial geometry data OR a simple features object that contains a geometry column}

\item{tolerance}{a value greater than 0}

\item{fileread}{Should the data be read from a file? Default is TRUE, use fileread=FALSE if the object is already a simple features object in the environment}
}
\value{
A data frame, with each row corresponding to a unique polygon point and its associated polygon and geographic information.
}
\description{
Get polygon and geographic information from a shapefile that is thinned to a more feasible dimension. This is team 12's results
}
\examples{
#get data for plotting
oz <- ozbig

aus <- team_12(file = oz, tolerance = 0.1, fileread=FALSE)

head(aus)

}
