% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_4.R
\name{team_4}
\alias{team_4}
\title{Get polygon and geographic information from a shapefile that is thinned to a more feasible dimension}
\usage{
team_4(file, tolerance = 0.1, fileread = TRUE)
}
\arguments{
\item{file}{path to a shapefile (.shp) with spatial geometry data OR a simple features object that contains a geometry column}

\item{tolerance}{a value greater than 0}

\item{fileread}{Should the data be read from a file? Default is TRUE, use fileread=FALSE if the object is already a simple features object in the environment}
}
\value{
A data frame, with each row corresponding to a unique polygon and its information. The last column ('data') is a dataframe with a set of points defining the extent of the polygon
}
\description{
Get polygon and geographic information from a shapefile that is thinned to a more feasible dimension
}
\examples{
#get data for plotting
oz <- data(australia)

aus <- team_4(file = oz, tolerance = 0.1, fileread=FALSE)

head(aus)

#the last column contains polygon geometry
aus$data[[1]]

}
