% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team1.R
\name{team_1}
\alias{team_1}
\title{lab3 work with team_1 function}
\usage{
team_1(file, tolerance, fileread = TRUE)
}
\arguments{
\item{file}{path to a shapefile (.shp) with spatial geometry data OR (if \code{fileread = F}) a simple features object}

\item{tolerance}{A numeric value greater than 0.}

\item{fileread}{If the file can be read in memory or not. Default is TRUE, use fileread=FALSE if the object is already a simple features object}
}
\value{
plot
}
\description{
lab3 work with team_1 function
}
\examples{
ozbig1 <- ozbig
team_1(file=ozbig1, tolerance=0.1, fileread=FALSE)
}
